Fixed all resolvers
