+++ src/queries/RootQuery.ts
@@ -13,8 +13,8 @@ import { User } from '../models/graphql/UserTypes';
 import { Tweet, TweetList } from '../models/graphql/TweetTypes';
 
 // RESOLVERS
-import { resolveUserDetails } from '../resolvers/UserSpecific';
-import { resolveTweet, resolveTweets } from '../resolvers/TweetSpecific';
+import UserResolver from '../resolvers/UserResolver';
+import TweetResolver from '../resolvers/TweetResolver';
 
 export const rootQuery = new GraphQLObjectType({
     name: 'Root',
@@ -30,7 +30,7 @@ export const rootQuery = new GraphQLObjectType({
                 userName: { type: GraphQLString },
                 id: { type: GraphQLString }
             },
-            resolve: (parent, args) => resolveUserDetails(args.userName, args.id)
+            resolve: (parent, args, context) => new UserResolver(context).resolveUserDetails(args.userName, args.id)
         },
         Tweet: {
             type: Tweet,
@@ -38,7 +38,7 @@ export const rootQuery = new GraphQLObjectType({
             args: {
                 id: { type: GraphQLString }
             },
-            resolve: (parent, args) => resolveTweet(args.id)
+            resolve: (parent, args, context) => new TweetResolver(context).resolveTweet(args.id)
         },
         Tweets: {
             type: TweetList,
@@ -55,7 +55,7 @@ export const rootQuery = new GraphQLObjectType({
                 count: { type: GraphQLInt, defaultValue: 20 },
                 cursor: { type: GraphQLString, defaultValue: '' }
             },
-            resolve: (parent, args) => resolveTweets(args)
+            resolve: (parent, args, context) => new TweetResolver(context).resolveTweets(args)
         }
     }
 })
